TwitterSignupでもUserFormオブジェクトを使用するよう変更した.
