+++ app/forms/user_form.rb
@@ -6,6 +6,8 @@ class UserForm
   attribute :password
   attribute :password_confirmation
   attribute :nickname
+  attribute :provider
+  attribute :uid
 
   VALID_PASSWORD_REGEX = /(?=.*?[a-z])(?=.*?[A-Z])[a-zA-Z\d!?_.$&%\-]{8,30}/
   validates :password, length: { in: 8..30 }, format: { with: VALID_PASSWORD_REGEX }
@@ -13,26 +15,40 @@ class UserForm
   validates :password_confirmation, presence: true
 
   VALID_EMAIL_REGEX = /[a-zA-Z0-9_.+-]+@([a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9]*\.)+[a-zA-Z]{2,}/
-  validates :email, presence: true, length: { maximum: 255 }, format: { with: VALID_EMAIL_REGEX } #, unless: :authentication
-  validate :validate_email_on_duplication
+  if @provider && @uid
+    validates :email, presence: true, length: { maximum: 255 }, format: { with: VALID_EMAIL_REGEX }
+    validate :validate_email_on_duplication
+  end
   validates :nickname, presence: true, length: { maximum: 30 }
 
   attr_accessor :games
 
   def save
+    @user = User.new(email: email&.downcase, password:, password_confirmation:, nickname:)
+    @user.build_authentication(provider: @provider, uid: @uid) if oauth_sigup?
+    games&.each_key { |game_id| @user.playing_games.build(game_id:) } if games.present?
+
     return false if invalid?
 
-    user = User.new(email: email.downcase, password:, password_confirmation:, nickname:)
-    games&.each_key { |game_id| user.playing_games.build(game_id:) } if games.present?
+    @user.save!
 
-    user.save!
+    @user.activate! if oauth_sigup?
 
     true
+  end
+
+  def build_authentication(provider:, uid:)
+    @provider = provider
+    @uid = uid
   end
 
   private
 
   def validate_email_on_duplication
-    errors.add(:email, :taken, value: email) if User.exists?(email: email.downcase)
+    errors.add(:email, :taken, value: email) if User.exists?(email: email&.downcase)
+  end
+
+  def oauth_sigup?
+    @provider && @uid
   end
 end
